Refactor file/folder hierarchy for client functions
All frontend clients (MongoDB, Sony PS Store, YouTube).
