+++ frontend/src/client/__tests__/mongo.test.js
@@ -1,4 +1,4 @@
-import Client from './Client';
+import mongo from '../mongo';
 const fetchMock = require('fetch-mock');
 
 const mockInfo = { info: 'this is simply a placeholder' };
@@ -8,7 +8,7 @@ describe('confirms routes and fetch methods for games collection in db', () => {
   it('returns a value after finding all games in db', () => {
     const maxResults = 25;
     fetchMock.get(`/games/find/all?max=${maxResults}`, { api: mockInfo });
-    return Client.findAllGames()
+    return mongo.findAllGames()
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -19,7 +19,7 @@ describe('confirms routes and fetch methods for games collection in db', () => {
   it('returns expected value after finding one game in db', () => {
     const id = 'UP9000-CUSA00552_00-THELASTOFUS00000';
     fetchMock.get(`/games/find/one/?id=${encodeURIComponent(id)}`, { api: mockInfo });
-    return Client.findOneGame(id)
+    return mongo.findOneGame(id)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -29,7 +29,7 @@ describe('confirms routes and fetch methods for games collection in db', () => {
 
   it('returns expected value after adding or updating a game in db', () => {
     fetchMock.post('/games/add', { api: mockInfo });
-    return Client.addOrUpdateGame(mockInfo)
+    return mongo.addOrUpdateGame(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -43,7 +43,7 @@ describe('confirms routes and fetch methods for price alerts collection in db', 
   it('returns expected value after creating a new price alert', () => {
     fetchMock.post('/priceAlerts/add', { api: mockInfo });
     expect.assertions(2);
-    return Client.createPriceAlert(mockInfo)
+    return mongo.createPriceAlert(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -54,7 +54,7 @@ describe('confirms routes and fetch methods for price alerts collection in db', 
   it('returns expected value after finding an existing price alert', () => {
     fetchMock.post('/priceAlerts/find/one', { api: mockInfo });
     expect.assertions(2);
-    return Client.findOnePriceAlert(mockInfo)
+    return mongo.findOnePriceAlert(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -65,7 +65,7 @@ describe('confirms routes and fetch methods for price alerts collection in db', 
   it('returns expected value deleting an existing price alert', () => {
     fetchMock.delete('/priceAlerts/delete', { api: mockInfo });
     expect.assertions(2);
-    return Client.deletePriceAlert(mockInfo)
+    return mongo.deletePriceAlert(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -79,7 +79,7 @@ describe('confirms routes and fetch methods for blacklist collection in db', () 
   it('returns expected value after checking if a user is on the email blacklist', () => {
     fetchMock.post('/blacklist/find/one', { api: mockInfo });
     expect.assertions(2);
-    return Client.checkBlacklist(mockInfo)
+    return mongo.checkBlacklist(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
@@ -90,7 +90,7 @@ describe('confirms routes and fetch methods for blacklist collection in db', () 
   it('returns expected value after adding a user to the email blacklist', () => {
     fetchMock.put('/blacklist/add', { api: mockInfo });
     expect.assertions(2);
-    return Client.addToBlacklist(mockInfo)
+    return mongo.addToBlacklist(mockInfo)
       .then(resp => {
         expect(resp).toBeDefined();
         expect(resp).toEqual(mockInfo);
