+++ frontend/src/redux/sagas.js
@@ -1,14 +1,14 @@
 import { all, call, put, takeLatest } from 'redux-saga/effects';
-import Client from '../Client';
-import sonyStore from '../sonyStore/sonyStore';
-import youTube from '../youTube/youTube';
+import mongo from '../client/mongo';
+import sony from '../client/sony';
+import youTube from '../client/youTube';
 import { showLoading, hideLoading } from 'react-redux-loading-bar';
 import * as types from './constants/actionTypes'
 
 
 export function* fetchAllGamesInDb(action) {
     try {
-        const allGames = yield call(Client.findAllGames, action.payload);
+        const allGames = yield call(mongo.findAllGames, action.payload);
         yield put({ type: types.FETCH_ALL_GAMES_IN_DB_SUCCEEDED, payload: allGames });
     } catch (e) {
         yield put({ type: types.FETCH_ALL_GAMES_IN_DB_FAILED, message: e.message });
@@ -17,7 +17,7 @@ export function* fetchAllGamesInDb(action) {
 
 export function* findPopularGames(action) {
     try {
-        const games = yield call(sonyStore.findPopularGames, action.payload);
+        const games = yield call(sony.findPopularGames, action.payload);
         yield put({ type: types.FIND_POPULAR_GAMES_SUCCEEDED, payload: games });
     } catch (e) {
         yield put({ type: types.FIND_POPULAR_GAMES_FAILED, message: e.message });
@@ -26,7 +26,7 @@ export function* findPopularGames(action) {
 
 export function* searchTitle(action) {
     try {
-        const searchResults = yield call(sonyStore.findGameByTitle, action.payload);
+        const searchResults = yield call(sony.findGameByTitle, action.payload);
         yield put({ type: types.SEARCH_BY_TITLE_SUCCEEDED, payload: searchResults });
     } catch (e) {
         yield put({ type: types.SEARCH_BY_TITLE_FAILED, message: e.message });
@@ -35,7 +35,7 @@ export function* searchTitle(action) {
 
 export function* generateAutoSuggestions(action) {
     try {
-        const autoSuggestions = yield call(sonyStore.findGameByTitle, action.payload.title, action.payload.maxResults);
+        const autoSuggestions = yield call(sony.findGameByTitle, action.payload.title, action.payload.maxResults);
         yield put({ type: types.GENERATE_AUTO_SUGGESTIONS_SUCCEEDED, payload: autoSuggestions });
     } catch (e) {
         yield put({ type: types.GENERATE_AUTO_SUGGESTIONS_FAILED, message: e.message });
@@ -45,7 +45,7 @@ export function* generateAutoSuggestions(action) {
 export function* makeActiveGame(action) {
     try {
         yield put(showLoading());
-        const activeGame = yield call(sonyStore.findGameById, action.payload);
+        const activeGame = yield call(sony.findGameById, action.payload);
         yield put({ type: types.MAKE_ACTIVE_GAME_SUCCEEDED, payload: activeGame });
     } catch (e) {
         yield put({ type: types.MAKE_ACTIVE_GAME_FAILED, message: e.message });
@@ -56,7 +56,7 @@ export function* makeActiveGame(action) {
 
 export function* submitPriceAlert(action) {
     try {
-        const priceAlert = yield call(Client.createPriceAlert, action.payload);
+        const priceAlert = yield call(mongo.createPriceAlert, action.payload);
         yield put({ type: types.SUBMIT_PRICE_ALERT_SUCCEEDED, payload: priceAlert });
     } catch (e) {
         yield put({ type: types.SUBMIT_PRICE_ALERT_FAILED, message: e.message });
@@ -65,8 +65,8 @@ export function* submitPriceAlert(action) {
 
 export function* fetchPriceAlert(action) {
     try {
-        const userInfo = yield call(Client.findOnePriceAlert, action.payload);
-        const activeGame = yield call(sonyStore.findGameById, userInfo.game_id);
+        const userInfo = yield call(mongo.findOnePriceAlert, action.payload);
+        const activeGame = yield call(sony.findGameById, userInfo.game_id);
         yield put({ type: types.FETCH_PRICE_ALERT_SUCCEEDED, payload: { activeGame, userInfo } });
     } catch (e) {
         yield put({ type: types.FETCH_PRICE_ALERT_FAILED, message: e.message });
@@ -75,7 +75,7 @@ export function* fetchPriceAlert(action) {
 
 export function* deletePriceAlert(action) {
     try {
-        const priceAlert = yield call(Client.deletePriceAlert, action.payload);
+        const priceAlert = yield call(mongo.deletePriceAlert, action.payload);
         yield put({ type: types.DELETE_PRICE_ALERT_SUCCEEDED, payload: priceAlert });
     } catch (e) {
         yield put({ type: types.DELETE_PRICE_ALERT_FAILED, message: e.message });
@@ -84,7 +84,7 @@ export function* deletePriceAlert(action) {
 
 export function* checkBlacklist(action) {
     try {
-        const blacklistInfo = yield call(Client.checkBlacklist, action.payload);
+        const blacklistInfo = yield call(mongo.checkBlacklist, action.payload);
         yield put({ type: types.CHECK_BLACKLIST_SUCCEEDED, payload: blacklistInfo });
     } catch (e) {
         yield put({ type: types.CHECK_BLACKLIST_FAILED, message: e.message });
@@ -93,7 +93,7 @@ export function* checkBlacklist(action) {
 
 export function* addToBlacklist(action) {
     try {
-        const blacklistInfo = yield call(Client.addToBlacklist, action.payload);
+        const blacklistInfo = yield call(mongo.addToBlacklist, action.payload);
         yield put({ type: types.ADD_TO_BLACKLIST_SUCCEEDED, payload: blacklistInfo });
     } catch (e) {
         yield put({ type: types.ADD_TO_BLACKLIST_FAILED, message: e.message });
