+++ frontend/src/redux/sagas.test.js
@@ -6,9 +6,9 @@ import { throwError } from 'redux-saga-test-plan/providers';
 import * as sagas from './sagas';
 import * as types from './constants/actionTypes';
 import * as actionCreators from './actions/actionCreators';
-import sonyStore from '../sonyStore/sonyStore';
-import Client from '../Client';
-import youTube from '../youTube/youTube';
+import sony from '../client/sony';
+import mongo from '../client/mongo';
+import youTube from '../client/youTube';
 
 
 const stubData = [{ placeholderData: 'some data' }];
@@ -19,7 +19,7 @@ describe('saga: fetchAllGamesInDb', () => {
   it('fetches all games in db', () => {
     return expectSaga(sagas.fetchAllGamesInDb, actionCreators.fetchAllGamesInDb())
       .provide([
-        [matchers.call.fn(Client.findAllGames), stubData]
+        [matchers.call.fn(mongo.findAllGames), stubData]
       ])
       .put({ type: types.FETCH_ALL_GAMES_IN_DB_SUCCEEDED, payload: stubData })
       .run();
@@ -27,7 +27,7 @@ describe('saga: fetchAllGamesInDb', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.fetchAllGamesInDb, actionCreators.fetchAllGamesInDb())
-      .provide([[matchers.call.fn(Client.findAllGames), throwError(error)]
+      .provide([[matchers.call.fn(mongo.findAllGames), throwError(error)]
       ])
       .put({ type: types.FETCH_ALL_GAMES_IN_DB_FAILED, message: error.message })
       .run();
@@ -35,10 +35,10 @@ describe('saga: fetchAllGamesInDb', () => {
 });
 
 describe('saga: findPopularGames', () => {
-  it('finds new games from sonyStore', () => {
+  it('finds new games from sony', () => {
     return expectSaga(sagas.findPopularGames, actionCreators.findPopularGames(25))
       .provide([
-        [matchers.call.fn(sonyStore.findPopularGames), stubData]
+        [matchers.call.fn(sony.findPopularGames), stubData]
       ])
       .put({ type: types.FIND_POPULAR_GAMES_SUCCEEDED, payload: stubData })
       .run();
@@ -46,7 +46,7 @@ describe('saga: findPopularGames', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.findPopularGames, actionCreators.findPopularGames(25))
-      .provide([[matchers.call.fn(sonyStore.findPopularGames), throwError(error)]
+      .provide([[matchers.call.fn(sony.findPopularGames), throwError(error)]
       ])
       .put({ type: types.FIND_POPULAR_GAMES_FAILED, message: error.message })
       .run();
@@ -57,7 +57,7 @@ describe('saga: searchTitle', () => {
   it('finds game by title', () => {
     return expectSaga(sagas.searchTitle, actionCreators.searchByTitle('A game title'))
       .provide([
-        [matchers.call.fn(sonyStore.findGameByTitle), stubData]
+        [matchers.call.fn(sony.findGameByTitle), stubData]
       ])
       .put({ type: types.SEARCH_BY_TITLE_SUCCEEDED, payload: stubData })
       .run();
@@ -65,7 +65,7 @@ describe('saga: searchTitle', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.searchTitle, actionCreators.searchByTitle('A game title'))
-      .provide([[matchers.call.fn(sonyStore.findGameByTitle), throwError(error)]
+      .provide([[matchers.call.fn(sony.findGameByTitle), throwError(error)]
       ])
       .put({ type: types.SEARCH_BY_TITLE_FAILED, message: error.message })
       .run();
@@ -76,7 +76,7 @@ describe('saga: generateAutoSuggestions', () => {
   it('generates autosuggestions by searching by title', () => {
     return expectSaga(sagas.generateAutoSuggestions, actionCreators.generateAutoSuggestions('A game title', 5))
       .provide([
-        [matchers.call.fn(sonyStore.findGameByTitle), stubData]
+        [matchers.call.fn(sony.findGameByTitle), stubData]
       ])
       .put({ type: types.GENERATE_AUTO_SUGGESTIONS_SUCCEEDED, payload: stubData })
       .run();
@@ -84,7 +84,7 @@ describe('saga: generateAutoSuggestions', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.generateAutoSuggestions, actionCreators.generateAutoSuggestions('A game title', 5))
-      .provide([[matchers.call.fn(sonyStore.findGameByTitle), throwError(error)]
+      .provide([[matchers.call.fn(sony.findGameByTitle), throwError(error)]
       ])
       .put({ type: types.GENERATE_AUTO_SUGGESTIONS_FAILED, message: error.message })
       .run();
@@ -95,7 +95,7 @@ describe('saga: makeActiveGame', () => {
   it('makes the selected game the `active game`', () => {
     return expectSaga(sagas.makeActiveGame, actionCreators.makeActiveGame('A game id'))
       .provide([
-        [matchers.call.fn(sonyStore.findGameById), stubData]
+        [matchers.call.fn(sony.findGameById), stubData]
       ])
       .put({ type: types.MAKE_ACTIVE_GAME_SUCCEEDED, payload: stubData })
       .run();
@@ -103,7 +103,7 @@ describe('saga: makeActiveGame', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.makeActiveGame, actionCreators.makeActiveGame('A game id'))
-      .provide([[matchers.call.fn(sonyStore.findGameById), throwError(error)]
+      .provide([[matchers.call.fn(sony.findGameById), throwError(error)]
       ])
       .put({ type: types.MAKE_ACTIVE_GAME_FAILED, message: error.message })
       .run();
@@ -114,7 +114,7 @@ describe('saga: submitPriceAlert', () => {
   it('creates a new price alert', () => {
     return expectSaga(sagas.submitPriceAlert, actionCreators.createPriceAlert(stubData))
       .provide([
-        [matchers.call.fn(Client.createPriceAlert), stubData]
+        [matchers.call.fn(mongo.createPriceAlert), stubData]
       ])
       .put({ type: types.SUBMIT_PRICE_ALERT_SUCCEEDED, payload: stubData })
       .run();
@@ -122,7 +122,7 @@ describe('saga: submitPriceAlert', () => {
 
   it('handles errors', () => {
     return expectSaga(sagas.submitPriceAlert, actionCreators.createPriceAlert(stubData))
-      .provide([[matchers.call.fn(Client.createPriceAlert), throwError(error)]
+      .provide([[matchers.call.fn(mongo.createPriceAlert), throwError(error)]
       ])
       .put({ type: types.SUBMIT_PRICE_ALERT_FAILED, message: error.message })
       .run();
@@ -133,8 +133,8 @@ describe('saga: fetchPriceAlert', () => {
   it('fetches an existing price alert', () => {
     return expectSaga(sagas.fetchPriceAlert, actionCreators.fetchPriceAlert(stubData))
       .provide([
-        [matchers.call.fn(Client.findOnePriceAlert), stubData],
-        [matchers.call.fn(sonyStore.findGameById), stubData]
+        [matchers.call.fn(mongo.findOnePriceAlert), stubData],
+        [matchers.call.fn(sony.findGameById), stubData]
       ])
       .put({ type: types.FETCH_PRICE_ALERT_SUCCEEDED, payload: { activeGame: stubData, userInfo: stubData } })
       .run();
@@ -143,7 +143,7 @@ describe('saga: fetchPriceAlert', () => {
   it('handles errors', () => {
     return expectSaga(sagas.fetchPriceAlert, actionCreators.fetchPriceAlert(stubData))
       .provide([
-        [matchers.call.fn(Client.findOnePriceAlert), throwError(error)]
+        [matchers.call.fn(mongo.findOnePriceAlert), throwError(error)]
       ])
       .put({ type: types.FETCH_PRICE_ALERT_FAILED, message: error.message })
       .run();
@@ -154,7 +154,7 @@ describe('saga: deletePriceAlert', () => {
   it('deletes an existing price alert', () => {
     return expectSaga(sagas.deletePriceAlert, actionCreators.deletePriceAlert(stubData))
       .provide([
-        [matchers.call.fn(Client.deletePriceAlert), stubData]
+        [matchers.call.fn(mongo.deletePriceAlert), stubData]
       ])
       .put({ type: types.DELETE_PRICE_ALERT_SUCCEEDED, payload: stubData })
       .run();
@@ -163,7 +163,7 @@ describe('saga: deletePriceAlert', () => {
   it('handles errors', () => {
     return expectSaga(sagas.deletePriceAlert, actionCreators.deletePriceAlert(stubData))
       .provide([
-        [matchers.call.fn(Client.deletePriceAlert), throwError(error)]
+        [matchers.call.fn(mongo.deletePriceAlert), throwError(error)]
       ])
       .put({ type: types.DELETE_PRICE_ALERT_FAILED, message: error.message })
       .run();
@@ -174,7 +174,7 @@ describe('saga: checkBlacklist', () => {
   it('checks blacklist for user email', () => {
     return expectSaga(sagas.checkBlacklist, actionCreators.checkBlacklist(stubData))
       .provide([
-        [matchers.call.fn(Client.checkBlacklist), stubData]
+        [matchers.call.fn(mongo.checkBlacklist), stubData]
       ])
       .put({ type: types.CHECK_BLACKLIST_SUCCEEDED, payload: stubData })
       .run();
@@ -183,7 +183,7 @@ describe('saga: checkBlacklist', () => {
   it('handles errors', () => {
     return expectSaga(sagas.checkBlacklist, actionCreators.checkBlacklist(stubData))
       .provide([
-        [matchers.call.fn(Client.checkBlacklist), throwError(error)]
+        [matchers.call.fn(mongo.checkBlacklist), throwError(error)]
       ])
       .put({ type: types.CHECK_BLACKLIST_FAILED, message: error.message })
       .run();
@@ -194,7 +194,7 @@ describe('saga: addToBlacklist', () => {
   it('adds user email to blacklist', () => {
     return expectSaga(sagas.addToBlacklist, actionCreators.addToBlacklist(stubData))
       .provide([
-        [matchers.call.fn(Client.addToBlacklist), stubData]
+        [matchers.call.fn(mongo.addToBlacklist), stubData]
       ])
       .put({ type: types.ADD_TO_BLACKLIST_SUCCEEDED, payload: stubData })
       .run();
@@ -203,7 +203,7 @@ describe('saga: addToBlacklist', () => {
   it('handles errors', () => {
     return expectSaga(sagas.addToBlacklist, actionCreators.addToBlacklist(stubData))
       .provide([
-        [matchers.call.fn(Client.addToBlacklist), throwError(error)]
+        [matchers.call.fn(mongo.addToBlacklist), throwError(error)]
       ])
       .put({ type: types.ADD_TO_BLACKLIST_FAILED, message: error.message })
       .run();
